No longer pushing GOT onto stack.
If/when we need this functionality, this should be done to ensure the
stack pointer remains 8 byte aligned. Things appear to work fine
when they're 4 byte aligned, except for varadic functions that
take doubles. Doubles are 8 bytes and the strd/ldrd instructions
that are used will do funny things with a 4 byte aligned stack pointer.

GOT (Global Offset Table) is used for position independent code that is
unable to use PC-relative addresses. My memory for when those instances
are escapes me.